bord/requirements.txt
python-dotenv==1.0.1
pillow==10.4.0